// src/app/rest/model/event-term-capacity-internal.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventTermCapacityInternalDTO { 
    eventTermId?: number;
    status?: EventTermCapacityInternalDTO.StatusEnum;
    capacity?: number;
    registrations?: number;
    registeredCount?: number;
    confirmedCount?: number;
}
export namespace EventTermCapacityInternalDTO {
    export type StatusEnum = 'CONCEPT' | 'QUEUE' | 'ACCEPTED' | 'CONFIRMED';
    export const StatusEnum = {
        Concept: 'CONCEPT' as StatusEnum,
        Queue: 'QUEUE' as StatusEnum,
        Accepted: 'ACCEPTED' as StatusEnum,
        Confirmed: 'CONFIRMED' as StatusEnum
    };
}